+++ src/Components/Core/Presentation/AvatarEditor/AvatarEditorPreview/AvatarEditorPreviewCamera/IAvatarEditorPreviewCameraPresenter.ts
@@ -1,0 +1,4 @@
+export default interface AvatarEditorPreviewCameraPresenter {
+  zoomInOnFace(): void;
+  zoomOutOnFace(): void;
+}